Struct windloading::WindLoads

pub struct WindLoads {
    pub loads: Vec<Option<Loads>>,
    pub time: Vec<f64>,
    // some fields omitted
}

Wind loads builder

This structure is used to read the forces and moments time series from a data file and to create the WindLoading structure

Fields
loads: Vec<Option<Loads>>

forces and moments time series

time: Vec<f64>

time vector

Implementations

impl WindLoads

pub fn from_pickle<P: AsRef<Path>>(path: P) -> Result<Self, WindLoadsError>

Reads the wind loads from a pickle file

pub fn range(self, t_min: f64, t_max: f64) -> Self

pub fn decimate(self, decimation_rate: usize) -> Self

pub fn n_sample(self, n_sample: usize) -> Result<Self, WindLoadsError>

Set the number of time sample

pub fn truss(self) -> Result<Self, WindLoadsError>

Selects loads on the truss

pub fn topend(self) -> Result<Self, WindLoadsError>

Selects loads on the top-end

pub fn m2_asm_topend(self) -> Result<Self, WindLoadsError>

pub fn cring(self) -> Result<Self, WindLoadsError>

Selects loads on the C-ring

pub fn gir(self) -> Result<Self, WindLoadsError>

Selects loads on the GIR

pub fn m1_cell(self) -> Result<Self, WindLoadsError>

Selects loads on the M1 cells

pub fn m1_segments(self) -> Result<Self, WindLoadsError>

Selects loads on the M1 segments

pub fn m2_segments(self) -> Result<Self, WindLoadsError>

Selects loads on the M2 segments

pub fn m2_asm_reference_bodies(self) -> Result<Self, WindLoadsError>

pub fn select_all(self) -> Result<Self, WindLoadsError>

Selects all loads

pub fn select_all_with_asm(self) -> Result<Self, WindLoadsError>

Selects all loads in the ASM configuration

pub fn build(self) -> Result<WindLoading, WindLoadsError>

Builds a wind loading source object